client/requirements.txt
Flask==1.1.2
flask-restx==1.3.0
Jinja2==2.11.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
libclang==16.0.6
MarkupSafe==2.0.1
protobuf==3.20.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
tensorflow==2.8.0
Werkzeug==0.16.0
itsdangerous==2.0.1
tensorflow-serving-api==2.8.0